[feat]hooksへ機能部分を移植
